feat(enums): refactor `ImageType` extension logic and cleanup
- Moved extension methods to partial class in `ImageType.cs` and removed `ImageTypeExtensions.cs`.
- Updated usages to new `GetValuesFast()` and cleaned up usings.
